+++ sdk/python/pulumi_aws/route53/__init__.py
@@ -13,4 +13,6 @@ from .resolver_rule_association import *
 from .zone import *
 from .zone_association import *
 from .get_delegation_set import *
+from .get_resolver_rule import *
+from .get_resolver_rules import *
 from .get_zone import *